Created login page and session capabilities
Created a Context Listener class, not yet implemented.

Created an index.html which is a login page. It doesn't authenticate the user but does create a new session that persists through the JSPs.

Note that there is a bug with the inputNewBug process in that it will fail if any of the text contains an apostrophe. This bug will be fixed