Changed edit bug process to prevent lost updates.
Before a Database processes an editBug request, it queries bugdb for the updated value of the bug being edited. If the values are not the same (between the original version the user edited and what is currently in the database, then the update operation is aborted and throws an exception.