Changed com.x460dot11.data.Database.updateBug method by syncronized.
Now, users can request edit pages as much as they want, but once they enter the method, it is locked so that no other thread can make any updates (editBug.jsp). Once it enters the method, there is a second check to query the existing bug record that the user started with versus what currently exists in the database. Thus, there is a double check to guard against lost updates.